Converted the settings widget module to use typescript
