+++ src/main/javascript/engine/models.ts
@@ -1,11 +1,11 @@
+import { EventType } from './enums/event-type';
 
 export interface Player extends Entity {
-    //TODO: Remove these from the game engine & replace here
-    getDialogs: () => any
-
-    getCombatSchedule: () => any
-
-    getEquipment: () => any
+	//TODO: Remove these from the game engine & replace here
+	getDialogs: () => any
+	getCombatSchedule: () => any
+	getEquipment: () => any
+	switchSheathing: () => any
 }
 
 export interface Npc extends Entity {
@@ -34,4 +34,14 @@ export interface DynamicMapSquare {
 
 export interface NodeHash {
 
+}
+
+export interface EventContext {
+	event: EventType;
+	trigger: number | string;
+	player: Player;
+	component?: number;
+	slot?: number;
+	fromslot?: number;
+	toslot?: number;
 }
