+++ src/main/javascript/modules/index.ts
@@ -1,0 +1,1 @@
+import './widget/settings';
