--- src/main/javascript/modules/widget/settings/bootstrap.js
@@ -1,21 +1,0 @@
-/**
- * Module to initialise the widget script bindings.
- */
-module.exports = function (scriptManager) {
-	var modules = [
-	    require('./combat-settings'),
-		require('./action-bar-settings'),
-		require('./doomsayer-settings'),
-		require('./game-settings'),
-		require('./gameplay-settings'),
-		require('./house-settings'),
-		require('./interface-settings'),
-		require('./loot-settings'),
-		require('./misc-settings'),
-		require('./player-inspect-settings')
-	];
-	
-	for (var i in modules) {
-		modules[i].init(scriptManager);
-	}
-};
