+++ src/main/javascript/modules/widget/settings/combat-settings.ts
@@ -19,33 +19,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-/* globals EventType */
-var util = require('shared/util');
+import { EventType } from 'engine/enums/event-type';
+import _events from 'engine/events';
+import { varbit, setVarBit } from 'engine/var';
+import { defaultHandler } from 'shared/util';
+import { sendMessage } from 'shared/chat';
 
-/**
- * @author Im Frizzy <skype:kfriz1998>
- * @author Frosty Teh Snowman <skype:travis.mccorkle>
- * @author Arthur <skype:arthur.behesnilian>
- * @author Kayla <skype:ashbysmith1996>
- * @author Sundays211
- * @since 08/01/2015
- */
-module.exports = (function () {
-	return {
-		init : init
-	};
-
-	function init (scriptManager) {
-		scriptManager.bind(EventType.IF_BUTTON, 970, function (ctx) {
-			switch (ctx.component) {
-			/*case 11://Secondary action bar
-			case 14://Tertiary action bar
-			case 18://Quaternary action bar
-			case 22://Quinary action bar*/
-			default:
-				util.defaultHandler(ctx, "action bar settings");
-				return;
-			}
-		});
+_events.bindEventListener(EventType.IF_BUTTON, 1503, (ctx) => {
+	var player = ctx.player;
+	switch (ctx.component) {
+		case 2://sheathe
+			player.switchSheathing();
+			return;
+		case 4://special attack
+			player.getCombatSchedule().updateAdrenaline(1000);
+			player.getCombatSchedule().setSpecialEnabled(!player.getCombatSchedule().isSpecialEnabled());
+			player.getCombatSchedule().setDefaultAttack();
+			sendMessage(ctx.player, "Special is " + (player.getCombatSchedule().isSpecialEnabled() ? "enabled." : "disabled."));
+			return;
+		case 49://retaliate
+			let wasRetaliating = varbit(ctx.player, 462) === 0;
+			setVarBit(ctx.player, 462, wasRetaliating ? 1 : 0);
+			return;
+		case 32://attack
+		case 36://balance
+		case 40://strength
+		case 44://defence
+			defaultHandler(ctx, "Combat Settings");
+			return;
+		default:
+			defaultHandler(ctx, "Combat Settings");
+			return;
 	}
-})();
+});
